requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
pydantic==2.7.4
pydantic-settings==2.3.4
sqlalchemy==2.0.31
psycopg[binary]==3.1.19
alembic==1.13.2
httpx==0.27.0
pinecone-client==3.2.2
openai
python-dotenv==1.0.1
langgraph>=0.2.0
langchain-openai>=0.2.0
langchain-core>=0.3.0
sse-starlette>=1.8.0
